app/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
influxdb-client==1.29.1
iniconfig==1.1.1
msgpack==1.0.4
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
Rx==3.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
